Added a k-d tree in an attempt to speed up performance.
This implementation is slower than the naive approach. I think this is because the naive approach is vectorized, yielding one subtraction and addition per test sample. The k-d tree approach can theoretically require n * m operations maximum for n training samples and m test samples. It is also a poor fit for digit recognition because the tree is unbalanced since most pixels are intensity 0.